Added CartController and logic (working)

+++ src/main/java/com/ibm/ojt/CartItem.java
@@ -3,7 +3,7 @@ package com.ibm.ojt;
 public class CartItem {
 	private String prodCode;
 	private int itemQty;
-	private int size;
+	private double size;
 	
 	public String getProdCode() {
 		return prodCode;
@@ -17,10 +17,10 @@ public class CartItem {
 	public void setItemQty(int itemQty) {
 		this.itemQty = itemQty;
 	}
-	public int getSize() {
+	public double getSize() {
 		return size;
 	}
-	public void setSize(int size) {
+	public void setSize(double size) {
 		this.size = size;
 	}
 }
